+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "read_c_vectors.h": "c"
+    }
+}
