+++ create_mat_demo/makefile
@@ -1,0 +1,2 @@
+createEpsilonMatrix: createEpsilonMatrix.c
+	gcc -o createEpsilonMatrix createEpsilonMatrix.c
