+++ create_mat_demo/output.txt
@@ -1,0 +1,18 @@
+Epsilon matrix:
+      e0001 e0002 e0003 e0004 e0010 e0020 e0030 e0040 e0100 e0200 e0300 e0400 e1000 e2000 e3000 e4000 
+e0001   3     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0002   0     3     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0003   0     0     3     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0004   0     0     0     3     0     0     0     0     0     0     0     0     0     0     0     0   
+e0010   0     0     0     0     3     0     0     0     0     0     0     0     0     0     0     0   
+e0020   0     0     0     0     0     3     0     0     0     0     0     0     0     0     0     0   
+e0030   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0040   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0100   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0200   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0300   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e0400   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e1000   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e2000   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e3000   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0   
+e4000   0     0     0     0     0     0     0     0     0     0     0     0     0     0     0     0
