+++ create_mat_demo/test.csv
@@ -1,0 +1,16 @@
+e0001, {0,0,0,1}
+e0002, {0,0,0,2}
+e0003, {0,0,0,3}
+e0004, {0,0,0,4}
+e0010, {0,0,1,0}
+e0020, {0,0,2,0}
+e0030, {0,0,3,0}
+e0040, {0,0,4,0}
+e0100, {0,1,0,0}
+e0200, {0,2,0,0}
+e0300, {0,3,0,0}
+e0400, {0,4,0,0}
+e1000, {1,0,0,0}
+e2000, {2,0,0,0}
+e3000, {3,0,0,0}
+e4000, {4,0,0,0}
