+++ demo/output/compressvector2.txt
@@ -1,0 +1,191 @@
+(1 1 2 1) : (1 0 1 3)
+(1 0 3 1) : (1 0 1 3)
+(1 0 2 2) : (1 0 1 3)
+(1 0 2 2) : (1 1 1 2)
+(1 0 2 2) : (0 2 1 2)
+(1 0 1 3) : (1 2 1 1)
+(1 0 1 3) : (1 3 0 1)
+(1 0 1 3) : (1 4 0 0)
+(1 1 1 2) : (1 4 0 0)
+(0 2 1 2) : (1 4 0 0)
+(1 2 1 1) : (0 5 0 0)
+(1 3 0 1) : (0 4 0 1)
+(1 4 0 0) : (0 3 0 2)
+(1 4 0 0) : (1 2 0 2)
+(1 4 0 0) : (1 1 1 2)
+(0 5 0 0) : (2 0 1 2)
+(0 4 0 1) : (3 0 1 1)
+(0 3 0 2) : (3 0 2 0)
+(1 2 0 2) : (2 1 2 0)
+(1 1 1 2) : (2 2 1 0)
+(2 0 1 2) : (1 2 1 1)
+(3 0 1 1) : (0 3 1 1)
+(3 0 2 0) : (0 4 0 1)
+(2 1 2 0) : (0 4 0 1)
+(2 2 1 0) : (0 3 1 1)
+(1 2 1 1) : (1 3 1 0)
+(0 3 1 1) : (1 3 1 0)
+(0 4 0 1) : (1 2 1 1)
+(0 4 0 1) : (1 2 1 1)
+(0 3 1 1) : (1 2 1 1)
+(1 3 1 0) : (0 2 1 2)
+(1 3 1 0) : (1 1 1 2)
+(1 2 1 1) : (2 1 1 1)
+(1 2 1 1) : (3 0 1 1)
+(1 2 1 1) : (3 0 0 2)
+(0 2 1 2) : (3 0 0 2)
+(1 1 1 2) : (2 0 1 2)
+(2 1 1 1) : (1 0 1 3)
+(3 0 1 1) : (0 0 1 4)
+(3 0 0 2) : (0 0 2 3)
+(3 0 0 2) : (0 1 2 2)
+(2 0 1 2) : (0 2 1 2)
+(1 0 1 3) : (0 2 1 2)
+(0 0 1 4) : (0 2 2 1)
+(0 0 2 3) : (0 2 2 1)
+(0 1 2 2) : (0 2 2 1)
+(0 2 1 2) : (1 1 2 1)
+(0 2 1 2) : (2 1 2 0)
+(0 2 2 1) : (2 1 1 1)
+(0 2 2 1) : (2 1 1 1)
+(0 2 2 1) : (2 1 1 1)
+(1 1 2 1) : (1 2 1 1)
+(2 1 2 0) : (0 2 1 2)
+(2 1 1 1) : (0 3 1 1)
+(2 1 1 1) : (0 3 0 2)
+(2 1 1 1) : (1 2 0 2)
+(1 2 1 1) : (1 1 1 2)
+(0 2 1 2) : (1 1 2 1)
+(0 3 1 1) : (2 0 2 1)
+(0 3 0 2) : (3 0 2 0)
+(1 2 0 2) : (3 0 2 0)
+(1 1 1 2) : (3 1 1 0)
+(1 1 2 1) : (3 1 0 1)
+(2 0 2 1) : (2 1 1 1)
+(3 0 2 0) : (1 1 1 2)
+(3 0 2 0) : (0 2 1 2)
+(3 1 1 0) : (0 2 1 2)
+(3 1 0 1) : (0 2 2 1)
+(2 1 1 1) : (0 2 1 2)
+(1 1 1 2) : (0 3 1 1)
+(0 2 1 2) : (1 2 1 1)
+(0 2 1 2) : (2 1 1 1)
+(0 2 2 1) : (2 1 0 2)
+(0 2 1 2) : (2 1 1 1)
+(0 3 1 1) : (2 0 2 1)
+(1 2 1 1) : (1 0 3 1)
+(2 1 1 1) : (0 1 3 1)
+(2 1 0 2) : (0 2 3 0)
+(2 1 1 1) : (1 2 2 0)
+(2 0 2 1) : (2 2 1 0)
+(1 0 3 1) : (3 2 0 0)
+(0 1 3 1) : (3 1 0 1)
+(0 2 3 0) : (3 1 0 1)
+(1 2 2 0) : (3 1 0 1)
+(2 2 1 0) : (2 1 1 1)
+(3 2 0 0) : (1 1 2 1)
+(3 1 0 1) : (1 1 3 0)
+(3 1 0 1) : (1 0 3 1)
+(3 1 0 1) : (0 0 4 1)
+(2 1 1 1) : (1 0 3 1)
+(1 1 2 1) : (2 0 2 1)
+(1 1 3 0) : (2 1 1 1)
+(1 0 3 1) : (2 2 1 0)
+(0 0 4 1) : (3 2 0 0)
+(1 0 3 1) : (2 3 0 0)
+(2 0 2 1) : (1 3 0 1)
+(2 1 1 1) : (1 3 0 1)
+(2 2 1 0) : (1 3 0 1)
+(3 2 0 0) : (1 3 0 1)
+(2 3 0 0) : (1 3 0 1)
+(1 3 0 1) : (1 4 0 0)
+(1 3 0 1) : (2 3 0 0)
+(1 3 0 1) : (3 2 0 0)
+(1 3 0 1) : (2 2 1 0)
+(1 3 0 1) : (2 1 1 1)
+(1 4 0 0) : (2 0 2 1)
+(2 3 0 0) : (1 0 3 1)
+(3 2 0 0) : (1 0 3 1)
+(2 2 1 0) : (1 0 2 2)
+(2 1 1 1) : (1 1 2 1)
+(2 0 2 1) : (1 1 1 2)
+(1 0 3 1) : (1 1 0 3)
+(1 0 3 1) : (0 1 0 4)
+(1 0 2 2) : (0 1 0 4)
+(1 1 2 1) : (0 1 0 4)
+(1 1 1 2) : (1 1 0 3)
+(1 1 0 3) : (1 2 0 2)
+(0 1 0 4) : (1 3 0 1)
+(0 1 0 4) : (1 3 0 1)
+(0 1 0 4) : (2 2 0 1)
+(1 1 0 3) : (1 2 0 2)
+(1 2 0 2) : (1 1 1 2)
+(1 3 0 1) : (1 0 2 2)
+(1 3 0 1) : (1 0 2 2)
+(2 2 0 1) : (0 0 2 3)
+(1 2 0 2) : (0 1 2 2)
+(1 1 1 2) : (1 1 1 2)
+(1 0 2 2) : (1 1 1 2)
+(1 0 2 2) : (1 1 2 1)
+(0 0 2 3) : (2 1 2 0)
+(0 1 2 2) : (3 0 2 0)
+(1 1 1 2) : (2 0 3 0)
+(1 1 1 2) : (3 0 2 0)
+(1 1 2 1) : (3 1 1 0)
+(2 1 2 0) : (3 1 1 0)
+(3 0 2 0) : (2 2 1 0)
+(2 0 3 0) : (3 2 0 0)
+(3 0 2 0) : (2 3 0 0)
+(3 1 1 0) : (3 2 0 0)
+(3 1 1 0) : (2 2 0 1)
+(2 2 1 0) : (3 1 0 1)
+(3 2 0 0) : (2 1 0 2)
+(2 3 0 0) : (2 0 0 3)
+(3 2 0 0) : (2 0 0 3)
+(2 2 0 1) : (2 0 0 3)
+(3 1 0 1) : (1 1 0 3)
+(2 1 0 2) : (1 1 0 3)
+(2 0 0 3) : (1 1 0 3)
+(2 0 0 3) : (0 1 0 4)
+(2 0 0 3) : (0 2 0 3)
+(1 1 0 3) : (0 1 1 3)
+(1 1 0 3) : (0 2 1 2)
+(1 1 0 3) : (0 2 1 2)
+(0 1 0 4) : (1 2 1 1)
+(0 2 0 3) : (1 1 2 1)
+(0 1 1 3) : (1 1 1 2)
+(0 2 1 2) : (1 0 1 3)
+(0 2 1 2) : (2 0 1 2)
+(1 2 1 1) : (1 1 1 2)
+(1 1 2 1) : (2 1 0 2)
+(1 1 1 2) : (3 1 0 1)
+(1 0 1 3) : (3 1 0 1)
+(2 0 1 2) : (2 1 0 2)
+(1 1 1 2) : (2 1 0 2)
+(2 1 0 2) : (1 2 0 2)
+(3 1 0 1) : (0 2 0 3)
+(3 1 0 1) : (0 2 0 3)
+(2 1 0 2) : (0 3 0 2)
+(2 1 0 2) : (0 2 1 2)
+(1 2 0 2) : (0 1 1 3)
+(0 2 0 3) : (0 2 1 2)
+(0 2 0 3) : (0 3 1 1)
+(0 3 0 2) : (1 2 1 1)
+(0 2 1 2) : (1 2 1 1)
+(0 1 1 3) : (1 3 1 0)
+(0 2 1 2) : (1 2 1 1)
+(0 3 1 1) : (1 1 2 1)
+(1 2 1 1) : (1 1 2 1)
+(1 2 1 1) : (1 1 2 1)
+(1 3 1 0) : (1 0 2 2)
+(1 2 1 1) : (1 1 2 1)
+(1 1 2 1) : (1 2 1 1)
+(1 1 2 1) : (1 2 1 1)
+(1 1 2 1) : (1 2 0 2)
+(1 0 2 2) : (1 3 0 1)
+(1 1 2 1) : (1 3 0 1)
+(1 2 1 1) : (2 2 0 1)
+(1 2 1 1) : (2 2 0 1)
+(1 2 0 2) : (2 2 1 0)
+(1 3 0 1) : (2 2 1 0)
+(1 3 0 1) : (2 1 2 0)
