--- output.txt
@@ -1,8 +1,0 @@
-Epsilon matrix:
-      e0001 e0002 e0001 e0002 e0010 e0020 
-e0001   0     2     0     0     0     0   
-e0002   0     0     0     0     0     0   
-e0001   0     0     0     0     0     0   
-e0002   0     0     0     0     0     0   
-e0010   0     0     0     0     0     1   
-e0020   0     0     0     0     0     0
